+++ src/main/java/com/ayakacraft/carpetayakaaddition/mixin/carpet/MinecraftServerMixin.java
@@ -18,22 +18,25 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.ayakacraft.carpetayakaaddition.utils;
+package com.ayakacraft.carpetayakaaddition.mixin.carpet;
 
-import com.ayakacraft.carpetayakaaddition.utils.preprocess.PreprocessPattern;
+import com.ayakacraft.carpetayakaaddition.CarpetAyakaServer;
+import com.ayakacraft.carpetayakaaddition.utils.mods.ModUtils;
+import me.fallenbreath.conditionalmixin.api.annotation.Condition;
+import me.fallenbreath.conditionalmixin.api.annotation.Restriction;
 import net.minecraft.server.MinecraftServer;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import java.nio.file.Path;
+@Restriction(require = @Condition(value = ModUtils.MC_ID, versionPredicates = "<1.16"))
+@Mixin(MinecraftServer.class)
+public class MinecraftServerMixin {
 
-public final class ServerUtils {
-
-    @PreprocessPattern
-    private static Path worldRootPath(MinecraftServer server) {
-        //#if MC>=11600
-        return server.getSavePath(net.minecraft.util.WorldSavePath.ROOT);
-        //#else
-        //$$ return server.getWorld(net.minecraft.world.dimension.DimensionType.OVERWORLD).getSaveHandler().getWorldDir().toPath();
-        //#endif
+    @Inject(method = "loadWorld", at = @At("TAIL"))
+    private void onLoadWorld(CallbackInfo ci) {
+        CarpetAyakaServer.INSTANCE.onServerLoadedWorlds$Ayaka();
     }
 
 }