+++ src/main/java/com/ayakacraft/carpetayakaaddition/utils/mixin/AyakaMixinConfigPlugin.java
@@ -23,7 +23,6 @@ package com.ayakacraft.carpetayakaaddition.utils.mixin;
 import com.ayakacraft.carpetayakaaddition.CarpetAyakaAddition;
 import me.fallenbreath.conditionalmixin.api.mixin.RestrictiveMixinConfigPlugin;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 
@@ -46,7 +45,7 @@ public class AyakaMixinConfigPlugin extends RestrictiveMixinConfigPlugin {
 
     @Override
     public List<String> getMixins() {
-        return new ArrayList<>();
+        return List.of();
     }
 
 }